[ethers-v4] Support for overloaded functions
