+++ packages/target-ethers-v4/src/codegen/index.ts
@@ -1,19 +1,14 @@
 import { values } from 'lodash'
 import {
-  AbiOutputParameter,
   AbiParameter,
   BytecodeWithLinkReferences,
   Contract,
-  FunctionDocumentation,
   EventArgDeclaration,
   EventDeclaration,
-  EvmOutputType,
-  EvmType,
   FunctionDeclaration,
-  isConstant,
-  isConstantFn,
-  TupleType,
 } from 'typechain'
+import { generateInputType, generateInputTypes } from './types'
+import { codegenFunctions } from './functions'
 
 export function codegenContractTypings(contract: Contract) {
   const template = `
@@ -52,16 +47,10 @@ export function codegenContractTypings(contract: Contract) {
     interface: ${contract.name}Interface;
 
     functions: {
-      ${values(contract.functions)
-        .map((v) => v[0])
-        .map(generateFunction)
-        .join('\n')}
+      ${values(contract.functions).map(codegenFunctions).join('\n')}
     };
 
-    ${values(contract.functions)
-      .map((v) => v[0])
-      .map(generateFunction)
-      .join('\n')}
+    ${values(contract.functions).map(codegenFunctions).join('\n')}
 
     filters: {
       ${values(contract.events)
@@ -199,19 +188,6 @@ function generateLibraryAddressesInterface(contract: Contract, bytecode: Bytecod
   };`
 }
 
-function generateFunction(fn: FunctionDeclaration): string {
-  return `
-  ${generateFunctionDocumentation(fn.documentation)}
-  ${fn.name}(${generateInputTypes(fn.inputs)}${
-    !isConstant(fn) && !isConstantFn(fn) ? 'overrides?: TransactionOverrides' : ''
-  }): Promise<${
-    fn.stateMutability === 'pure' || fn.stateMutability === 'view'
-      ? generateOutputTypes(fn.outputs)
-      : 'ContractTransaction'
-  }>;
-`
-}
-
 function generateEstimateFunction(fn: FunctionDeclaration): string {
   return `
   ${fn.name}(${generateInputTypes(fn.inputs)}): Promise<BigNumber>;
@@ -224,47 +200,6 @@ function generateInterfaceFunctionDescription(fn: FunctionDeclaration): string {
     fn.inputs,
   )}): string; }>;
 `
-}
-
-function generateFunctionDocumentation(doc?: FunctionDocumentation): string {
-  if (!doc) return ''
-
-  let docString = '/**'
-  if (doc.details) docString += `\n * ${doc.details}`
-  if (doc.notice) docString += `\n * ${doc.notice}`
-
-  const params = Object.entries(doc.params || {})
-  if (params.length) {
-    params.forEach(([key, value]) => {
-      docString += `\n * @param ${key} ${value}`
-    })
-  }
-
-  if (doc.return) docString += `\n * @returns ${doc.return}`
-
-  docString += '\n */'
-
-  return docString
-}
-
-function generateInputTypes(input: Array<AbiParameter>): string {
-  if (input.length === 0) {
-    return ''
-  }
-  return (
-    input.map((input, index) => `${input.name || `arg${index}`}: ${generateInputType(input.type)}`).join(', ') + ', '
-  )
-}
-
-function generateOutputTypes(outputs: Array<AbiOutputParameter>): string {
-  if (outputs.length === 1) {
-    return generateOutputType(outputs[0].type)
-  } else {
-    return `{
-      ${outputs.map((t) => t.name && `${t.name}: ${generateOutputType(t.type)}, `).join('')}
-      ${outputs.map((t, i) => `${i}: ${generateOutputType(t.type)}`).join(', ')}
-      }`
-  }
 }
 
 function generateParamArrayTypes(params: Array<AbiParameter>): string {
@@ -312,63 +247,4 @@ function generateEventTypes(eventArgs: EventArgDeclaration[]) {
 
 function generateEventArgType(eventArg: EventArgDeclaration): string {
   return eventArg.isIndexed ? `${generateInputType(eventArg.type)} | null` : 'null'
-}
-
-// https://docs.ethers.io/ethers.js/html/api-contract.html#types
-function generateInputType(evmType: EvmType): string {
-  switch (evmType.type) {
-    case 'integer':
-      return 'BigNumberish'
-    case 'uinteger':
-      return 'BigNumberish'
-    case 'address':
-      return 'string'
-    case 'bytes':
-    case 'dynamic-bytes':
-      return 'Arrayish'
-    case 'array':
-      return `(${generateInputType(evmType.itemType)})[]`
-    case 'boolean':
-      return 'boolean'
-    case 'string':
-      return 'string'
-    case 'tuple':
-      return generateTupleType(evmType, generateInputType)
-  }
-}
-
-function generateOutputType(evmType: EvmOutputType): string {
-  switch (evmType.type) {
-    case 'integer':
-    case 'uinteger':
-      return evmType.bits <= 48 ? 'number' : 'BigNumber'
-    case 'address':
-      return 'string'
-    case 'void':
-      return 'void'
-    case 'bytes':
-    case 'dynamic-bytes':
-      return 'string'
-    case 'array':
-      return `(${generateOutputType(evmType.itemType)})[]`
-    case 'boolean':
-      return 'boolean'
-    case 'string':
-      return 'string'
-    case 'tuple':
-      return generateOutputTupleType(evmType)
-  }
-}
-
-function generateTupleType(tuple: TupleType, generator: (evmType: EvmType) => string) {
-  return '{' + tuple.components.map((component) => `${component.name}: ${generator(component.type)}`).join(',') + '}'
-}
-
-function generateOutputTupleType(tuple: TupleType) {
-  return (
-    '{' +
-    tuple.components.map((component) => `${component.name}: ${generateOutputType(component.type)} ,`).join('\n') +
-    tuple.components.map((component, index) => `${index}: ${generateOutputType(component.type)}`).join(', ') +
-    '}'
-  )
 }
